Détection de maladie/fastapi_resnet/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pillow==11.3.0
numpy==2.3.2
onnxruntime==1.24.4
python-multipart==0.0.20
